Failing tests for reversing parse with clean_node
